+++ src/tilda/loader/csv/ImportProcessor.java
@@ -16,7 +16,6 @@
 
 package tilda.loader.csv;
 
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
@@ -29,8 +28,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import tilda.loader.csv.ImporterThread;
-import tilda.loader.csv.CSVImporter.Results;
-import tilda.loader.parser.Config;
+import tilda.loader.csv.stores.CSVImporter.Results;
 import tilda.loader.parser.DataObject;
 
 import tilda.db.Connection;
@@ -39,51 +37,6 @@ import tilda.utils.DurationUtil;
 public class ImportProcessor
   {
     protected static final Logger LOG = LogManager.getLogger(ImportProcessor.class.getName());
-
-    public static void process(Connection C, Config Conf, List<DataObject> CMSDataList)
-      {
-        try
-          {
-            List<Results> Results = new ArrayList<Results>();
-            for (DataObject Data : CMSDataList)
-              {
-                validate(Data);
-                List<Results> Res = CSVImporter.process(C, Conf, Data);
-                if (Res == null)
-                  break;
-                Results.addAll(Res);
-              }
-
-            long totalCount = 0;
-            long totalNano = 0;
-            for (Results R : Results)
-              {
-                totalCount += R._RecordsCount;
-                totalNano += R._TimeNano;
-                LOG.debug("Processed file " + R._FileName + " into table " + R._TableName + " in " + DurationUtil.PrintDurationSeconds(R._TimeNano) +
-                " (" + DurationUtil.PrintPerformancePerMinute(R._TimeNano, R._RecordsCount) + " Records/min)");
-              }
-            LOG.debug("--------------------------------------------------------------------------------------------------------------");
-            LOG.debug("In total, processed " + totalCount + " in " + DurationUtil.PrintDuration(totalNano) + " (" + DurationUtil.PrintPerformancePerMinute(totalNano, totalCount) + " Records/min)");
-
-          }
-        catch (Exception E)
-          {
-            LOG.error("Error: ", E);
-          }
-        finally
-          {
-            try
-              {
-                if (C != null)
-                  C.close();
-              }
-            catch (SQLException e)
-              {
-                LOG.error("Error in the application: ", e);
-              }
-          }
-      }
 
     /*
      * Launch and Shutdown threads
